Configure linting and enable in build step
This commit switches to the ESLint recommended rules, but disables two
of them so that the build can pass. **These should be re-enabled and
the errors fixed (see the root `.eslintrc`).** Note that the
configurations for `src/` and `BE/` are slightly different, to reflect
that one is React browser code and one is Node server code.

It also adds two NPM commands, `lint` and `lint:fix` (which runs the
linter and automatically fixes some simple issues).